requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
neo4j==5.18.0
numpy==1.26.4
openai==1.13.3
pandas==2.2.1
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1
